+++ some-api-client/src/main/java/org/spring/boot/metrics/instances/some/api/client/service/CreateService.java
@@ -17,5 +17,5 @@ public interface CreateService<S, Q> {
      * @param detail the detail
      * @return the s
      */
-    S create(Q detail);
+    S create(Q detail) throws Exception;
 }